Move getId out of Model.ts to a util file
Previously getId was being imported by the demo apps causing it to error.
